backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.11.0",
]
